requirements.txt
easyocr>=1.7.1
nba_api>=1.4.1
pandas>=2.1.0
numpy>=1.24.0
torch>=2.1.0
gspread>=5.12.0
google-auth>=2.23.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
python-dotenv>=1.0.0
requests>=2.31.0
SQLAlchemy>=2.0.0
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0 